pyproject.toml
# =============================================================================
# Author : B.Vignesh Kumar aka Bravetux <ic19939@gmail.com>
# Date   : 26 March 2026
# =============================================================================
[project]
name = "techtrainer-ai"
version = "0.1.0"
description = "AI-powered employee training chatbot built on Strands Agents"
requires-python = ">=3.10"
dependencies = [
    "strands-agents>=0.1.7",
    "strands-agents-tools>=0.1.7",
    "strands-agents-evals>=0.1.0",
    "streamlit>=1.35.0",
    "chromadb>=0.5.0",
    "boto3>=1.34.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
    "openpyxl>=3.1.0",
    "langchain-text-splitters>=0.2.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "litellm>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
